v0.25
Now supports USB!
2 more Homebrews added
(VTSPS2-EJECT & VTSPS2-TESTMODE)
Now supports 17 total Homebrews
No longer in ALPHA, First BETA release.
