+++ VTSPS2-HBDL.TXT
@@ -1,5 +1,5 @@
-;v0.23
-;2008-01-04 CB10.ELF v10
+;v0.25
+;2020-06-05 EJECT.ELF v0.6
 ;2011-02-02 ESDL.ELF v0.810 BETA
 ;2008-09-05 ESR.ELF R9B DVD Direct
 ;2019-04-16 GSM.ELF v0.38
@@ -14,8 +14,9 @@
 ;2020-05-16 RA_QNES.ELF RetroArch v1.8.8 QuickNES
 ;2019-05-06 SMS.ELF v2.9 REV4
 ;2016-11-23 SNESSTN.ELF v0.26C
+;2020-06-05 TESTMODE.ELF v0.3
 ;2019-04-19 WLE.ELF v4.43A
-CB10.ELF FB88EC2B
+EJECT.ELF 0B9903A0
 ESDL.ELF 8BDC61A0
 ESR.ELF 9EAE97B4
 GSM.ELF 226DA7F1
@@ -30,4 +31,5 @@ RA_PICO.ELF 9D421180
 RA_QNES.ELF B99079E0
 SMS.ELF 453A409F
 SNESSTN.ELF 7D5AFAF2
+TESTMODE.ELF 13D238E0
 WLE.ELF FFEAB9E9
