+++ ps2ipc.c
@@ -9,7 +9,7 @@
 #
 */
 
-#include "main.h"
+#include "VTSPS2-HBDL.h"
